Move `compactRange` implementation to `lib/state`
