+++ lib/state.js
@@ -2,11 +2,13 @@ const ReadyResource = require('ready-resource')
 const RefCounter = require('refcounter')
 const rrp = require('resolve-reject-promise')
 const SignalPromise = require('signal-promise')
+const c = require('compact-encoding')
 const { ReadBatch, WriteBatch } = require('./batch')
 const ColumnFamily = require('./column-family')
 const binding = require('../binding')
 
 const MAX_BATCH_REUSE = 64
+const empty = Buffer.alloc(0)
 
 module.exports = class RocksDBState extends ReadyResource {
   constructor(db, path, opts) {
@@ -308,5 +310,48 @@ module.exports = class RocksDBState extends ReadyResource {
       if (err) req.reject(new Error(err))
       else req.resolve()
     }
+  }
+
+  async compactRange(db, start, end, opts) {
+    if (this.opened === false) await this.ready()
+
+    this.io.inc()
+
+    const { exclusive = false } = opts
+
+    const req = { resolve: null, reject: null, handle: null }
+
+    const promise = new Promise((resolve, reject) => {
+      req.resolve = resolve
+      req.reject = reject
+    })
+
+    try {
+      req.handle = binding.compactRange(
+        this._handle,
+        db._columnFamily._handle,
+        this._encodeKey(start),
+        this._encodeKey(end),
+        exclusive,
+        req,
+        oncompactrange
+      )
+
+      await promise
+    } finally {
+      this.io.dec()
+    }
+
+    function oncompactrange(err) {
+      if (err) req.reject(new Error(err))
+      else req.resolve()
+    }
+  }
+
+  _encodeKey(k) {
+    if (this.db._keyEncoding) return c.encode(this.db._keyEncoding, k)
+    if (typeof k === 'string') return Buffer.from(k)
+    if (k === null) return empty
+    return k
   }
 }
